좌석 재설정도 조건 추가

+++ khFinalproject/src/main/java/com/kh/spring/service/SeatServiceImpl.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
 import com.kh.spring.entity.theater.SeatDto;
+import com.kh.spring.repository.schedule.ScheduleTimeDao;
 import com.kh.spring.repository.theater.HallDao;
 import com.kh.spring.repository.theater.SeatDao;
 
@@ -19,24 +20,27 @@ public class SeatServiceImpl implements SeatService{
 	private HallDao hallDao;
 	@Autowired
 	private SeatDao seatDao;
+	@Autowired
+	private ScheduleTimeDao scheduleTimeDao;
 	
 	@Override
 	public void setSeatandUpdateHall(List<SeatDto> seatList, int hallNo) {
 		
-		seatDao.delete(hallNo); //일단 좌석 삭제하고 다시 생성
+		if(scheduleTimeDao.isScheduleTimeExist(hallNo)) { //해당 상영관에 상영 스케줄이 없어야 좌석을 만듦 
+			seatDao.delete(hallNo); //일단 좌석 삭제하고 다시 생성
 
-		int count = 0;
-		for(SeatDto seatDto : seatList) {
-			seatDto.setHallNo(hallNo);
-			seatDao.insert(seatDto);
-			
-			if(seatDto.getSeatStatus().contains("normal")) {
-				count++;
+			int count = 0;
+			for(SeatDto seatDto : seatList) {
+				seatDto.setHallNo(hallNo);
+				seatDao.insert(seatDto);
+				
+				if(seatDto.getSeatStatus().contains("normal")) {
+					count++;
+				}
 			}
+			
+			//hall에 총 좌석수를 업데이트 해준다.
+			hallDao.update(hallNo,count);
 		}
-		
-		//hall에 총 좌석수를 업데이트 해준다.
-		hallDao.update(hallNo,count);
 	}
-
 }
